Refactoring:
* Refactoring analytic routes: separate API route and View route
* Refactoring analytic views: add report view, analytics view(index) and management tool view
