+++ public/javascripts/consoles/analytics/report-controller.js
@@ -1,0 +1,14 @@
+/**
+ * Created by Roman Lo on 5/18/2016.
+ */
+
+define([
+    'angular',
+    'webapp',
+    'models/analytics-field',
+    'models/analytics-compare-set',
+    'models/analytics-field-filter',
+    'services/log-analytics-service',
+], function (angular, webapp, AnalyticsField, AnalyticsCompareSet, AnalyticsFieldFilter) {
+
+})
