+++ views/includes/analytics/search-panel.jade
@@ -4,7 +4,7 @@
 
 div.row.search-option-section
     div.col-md-12
-        div.panel
+        div.panel.panel-default
             div.panel-body
                 .form-inline
                     //- From Date Options
