+++ views/includes/nav-banner.jade
@@ -16,8 +16,17 @@ nav.navbar.navbar-default.navbar-fixed-top.navbar-inverse
         ul.nav.navbar-nav
           li
             a(href="/aliyun-sls/") SLS Console
-          li
-            a(href='/aliyun-sls-analytics/') Analytics
+          li.dropdown
+            a.dropdown-toggle(href="#", data-toggle="dropdown") Analytics&nbsp;
+              span.caret
+            ul.dropdown-menu
+              li(role="menuitem")
+                a(href="/analytics/") CompareSet
+              li(role="menuitem")
+                a(href="/analytics/report") Report
+              li.divider(role="separator")
+              li(role="menuitem")
+                a(href="/analytics/management") Management Tool
 
       ul.nav.navbar-nav.navbar-right
         if (user)
